analyzer/requirements.txt
ffmpeg-python
requests
python-dotenv
